Separate interpreter selectors into `implementation` and `platform` modules
# Conflicts:
#	crates/uv-interpreter/src/lib.rs
