+++ crates/uv-interpreter/src/lib.rs
@@ -25,11 +25,12 @@ pub use crate::target::Target;
 
 mod environment;
 mod find_python;
+mod implementation;
 mod interpreter;
 pub mod managed;
+pub mod platform;
 mod py_launcher;
 mod python_version;
-pub mod selectors;
 mod target;
 
 #[derive(Debug, Error)]
